requirements.txt
flask==3.0.0
flask-socketio==5.3.6
opencv-python==4.9.0.80
ultralytics==8.1.0
numpy==1.26.3
python-dotenv==1.0.0
eventlet==0.35.1
